requirements.txt
numpy==1.11.1
Pillow==3.3.1
PyOpenGL==3.1.0
PyOpenGL-accelerate==3.1.0
